Rename `stateful::agent::state` to `field`
